test terminate all instances

+++ .github/scripts/TerminateTestInstances.sh
@@ -6,5 +6,9 @@ Instances=$(aws ec2 describe-instances \
 --filters Name=tag:Name,Values=Test_Quote_Server Name=instance-state-name,Values=running,pending,stopped,stopping \
 --output=text)
 
-echo $Instances
-aws ec2 terminate-instances --instance-ids $Instances
+if [ x"${Instances}" == "x" ]; then
+     echo "No value found for the variable"
+  else
+     echo "Value found for the variable $Instances"
+     aws ec2 terminate-instances --instance-ids $Instances
+  fi
